Added comments and removed unnecessary function
Added comments to the hamming subroutines so that doing the write up
is a bit easier. Also while looking at the code I realized that the
program does not need to lookup the syndrome index in the generator
matrix, we already have i when iterating through the generator matrix.
This makes the function hamming_get_syndrome_index() unnecessary.